requirements.txt
amuse-framework
amuse-seba
amuse-sse
amuse-bhtree
amuse-fi
natsort
numpy
matplotlib
plotly
